debug DSP-ADPCM decoding

+++ src/dspadpcmcodec.h
@@ -13,8 +13,8 @@ public:
     std::int32_t loopStart;
     std::uint32_t loopEnd;
     float gain;
-    std::uint16_t history[2];
-    std::uint16_t coefs[16];
+    std::int16_t history[2];
+    std::int16_t coefs[16];
   };
 
   DspAdpcmCodec(ClefContext* ctx, const Params& params);
@@ -22,7 +22,7 @@ public:
   virtual SampleData* decodeRange(std::vector<uint8_t>::const_iterator start, std::vector<uint8_t>::const_iterator end, uint64_t sampleID = SampleData::Uncached);
 
 private:
-  std::int16_t getNextSample(std::int8_t state);
+  std::int16_t getNextSample(std::int32_t state);
 
   Params params;
   std::int32_t scale, coef1, coef2;
